# swagger.yaml
openapi: 3.0.0
info:
  title: Events API
  description: API for searching events within a specified time range.
  version: 1.0.0
paths:
  /search:
    get:
      summary: Lists the available events on a time range
      operationId: searchEvents
      parameters:
        - name: starts_at
          in: query
          required: true
          description: Return only events that start after this date
          schema:
            type: string
            format: date-time
        - name: ends_at
          in: query
          required: true
          description: Return only events that finish before this date
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: List of events
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      events:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                              format: uuid
                            title:
                              type: string
                            start_date:
                              type: string
                              format: date
                            start_time:
                              type: string
                              format: time
                            end_date:
                              type: string
                              format: date
                            end_time:
                              type: string
                              format: time
                            min_price:
                              type: number
                              format: float
                            max_price:
                              type: number
                              format: float
                  error:
                    type: object
                    nullable: true
        '400':
          description: The request was not correctly formed (missing required parameters, wrong types...)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                  data:
                    type: object
                    nullable: true
        '500':
          description: Generic error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                  data:
                    type: object
                    nullable: true